Add Xcode format and readme
This commit adds a `--xcode` option to the tool that outputs errors in this
format:

    file:line: error: message

That format is picked up by Xcode and errors are displayed right in the UI,
allowing you to click on the error and jump directly to the offending line
in the project file.

Also added a Readme.